+++ framework/Furion/App/Extensions/AppServiceCollectionExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using Furion;
 using Furion.DependencyInjection;
 using Furion.UnifyResult;
+using Microsoft.Extensions.Hosting;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using System;
 using System.Linq;
@@ -144,6 +145,27 @@ namespace Microsoft.Extensions.DependencyInjection
         }
 
         /// <summary>
+        /// 自动添加主机服务
+        /// </summary>
+        /// <param name="services"></param>
+        /// <returns></returns>
+        public static IServiceCollection AddAppHostedService(this IServiceCollection services)
+        {
+            // 获取所有 BackgroundService 类型
+            var backgroundServiceTypes = App.EffectiveTypes.Where(u => typeof(BackgroundService).IsAssignableFrom(u));
+            var addHostServiceMethod = typeof(ServiceCollectionHostedServiceExtensions).GetMethods(BindingFlags.Static | BindingFlags.Public)
+                                .Where(u => u.Name.Equals("AddHostedService") && u.IsGenericMethod && u.GetParameters().Length == 1)
+                                .FirstOrDefault();
+
+            foreach (var type in backgroundServiceTypes)
+            {
+                addHostServiceMethod.MakeGenericMethod(type).Invoke(null, new object[] { services });
+            }
+
+            return services;
+        }
+
+        /// <summary>
         /// 添加应用配置
         /// </summary>
         /// <param name="services">服务集合</param>
@@ -178,6 +200,39 @@ namespace Microsoft.Extensions.DependencyInjection
 
             // 添加对象映射
             services.AddObjectMapper();
+
+            // 自定义服务
+            configure?.Invoke(services);
+
+            return services;
+        }
+
+        /// <summary>
+        /// 添加主机应用配置
+        /// </summary>
+        /// <param name="services">服务集合</param>
+        /// <param name="configure">服务配置</param>
+        /// <returns>服务集合</returns>
+        internal static IServiceCollection AddHostApp(this IServiceCollection services, Action<IServiceCollection> configure = null)
+        {
+            // 注册内存和分布式内存
+            services.AddMemoryCache();  // .NET 5.0.3+ 需要手动注册了
+            services.AddDistributedMemoryCache();
+
+            // 注册全局配置选项
+            services.AddConfigurableOptions<AppSettingsOptions>();
+
+            // 注册全局 Startup 扫描
+            services.AddStartup();
+
+            // 注册全局依赖注入
+            services.AddDependencyInjection();
+
+            // 添加对象映射
+            services.AddObjectMapper();
+
+            // 添加主机服务
+            services.AddAppHostedService();
 
             // 自定义服务
             configure?.Invoke(services);